Move the 'Hello Node' test code so we can make something useful
