+++ src/rustTest.js
@@ -1,9 +1,4 @@
-const ffi = require('ffi');
+// const ffi = require('ffi');
 
-const nativeRustFFI = ffi.Library('rust/target/release/libdecnalab', {
-    hello_from_rust: ['string', ['string']]
-});
-
-const string = nativeRustFFI.hello_from_rust('Node');
-
-console.log(string);
+// const nativeRustFFI = ffi.Library('rust/target/release/libdecnalab', {
+// });
